+++ tests/Fixture/example-config.php
@@ -38,16 +38,13 @@ $CFG->behat_wwwroot       = 'http://localhost:8000';
 $CFG->behat_faildump_path = '/path/to/moodledata/behat_dump';
 $CFG->behat_profiles      = [
     'default' => [
-        'browser'    => 'firefox',
-        'wd_host'    => 'http://localhost:4444/wd/hub',
-        'extensions' => [
-            'Behat\MinkExtension' => [
-                'selenium2' => [
-                    'browser' => 'firefox',
-                ]
-            ]
-        ]
-    ]
+        'browser' => 'firefox',
+        'wd_host' => 'http://localhost:4444/wd/hub',
+    ],
+    'chrome' => [
+        'browser' => 'chrome',
+        'wd_host' => 'http://localhost:4444/wd/hub',
+    ],
 ];
 
 // Extra config.