+++ docs/CHANGELOG.md
@@ -9,6 +9,9 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 The format of this change log follows the advice given at [Keep a CHANGELOG](http://keepachangelog.com).
 
 ## [Unreleased]
+No unreleased changes.
+
+## [2.1.0] - 2017-09-13
 ### Fixed
 - ACTION REQUIRED: added `firefox: "47.0.1"` to `.travis.dist.yml` and you must add it to your
   `.travis.yml` file to get Behat running again.  This is because Travis CI changed their default
@@ -187,7 +190,8 @@ The format of this change log follows the advice given at [Keep a CHANGELOG](htt
 - `moodle-plugin-ci shifter` command.  Run YUI Shifter on plugin YUI modules.
 - `moodle-plugin-ci csslint` command.  Lints the CSS files in the plugin.
 
-[Unreleased]: https://github.com/moodlerooms/moodle-plugin-ci/compare/2.0.1...master
+[Unreleased]: https://github.com/moodlerooms/moodle-plugin-ci/compare/2.1.0...master
+[2.1.0]: https://github.com/moodlerooms/moodle-plugin-ci/compare/2.0.1...2.1.0
 [2.0.1]: https://github.com/moodlerooms/moodle-plugin-ci/compare/2.0.0...2.0.1
 [2.0.0]: https://github.com/moodlerooms/moodle-plugin-ci/compare/1.5.8...2.0.0
 [1.5.8]: https://github.com/moodlerooms/moodle-plugin-ci/compare/1.5.7...1.5.8